Move copy_emoji_into_server to main.py
